moved details pannel to left and css addtition
